requirements.txt
certifi==2021.5.30
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
torch==1.9.0
torchsummary==1.5.1
torchvision==0.10.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
